+++ src/Illuminate/Image/Drivers/InterventionDriver.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Illuminate\Foundation\Image\Drivers;
+namespace Illuminate\Image\Drivers;
 
 use finfo;
 use Illuminate\Contracts\Image\Driver;
-use Illuminate\Foundation\Image\ImageException;
-use Illuminate\Foundation\Image\PendingImageOptions;
+use Illuminate\Image\ImageException;
+use Illuminate\Image\PendingImageOptions;
 use Intervention\Image\Encoders\JpegEncoder;
 use Intervention\Image\Encoders\MediaTypeEncoder;
 use Intervention\Image\Encoders\WebpEncoder;
