+++ src/Illuminate/Image/PendingImageOptions.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Illuminate\Foundation\Image;
+namespace Illuminate\Image;
 
 class PendingImageOptions
 {
