+++ src/Illuminate/Support/DefaultProviders.php
@@ -29,7 +29,7 @@ class DefaultProviders
             \Illuminate\Database\DatabaseServiceProvider::class,
             \Illuminate\Encryption\EncryptionServiceProvider::class,
             \Illuminate\Filesystem\FilesystemServiceProvider::class,
-            \Illuminate\Foundation\Image\ImageServiceProvider::class,
+            \Illuminate\Image\ImageServiceProvider::class,
             \Illuminate\Foundation\Providers\FoundationServiceProvider::class,
             \Illuminate\Hashing\HashServiceProvider::class,
             \Illuminate\Mail\MailServiceProvider::class,
