+++ src/Illuminate/Support/Facades/Image.php
@@ -2,16 +2,16 @@
 
 namespace Illuminate\Support\Facades;
 
-use Illuminate\Foundation\Image\ImageManager;
+use Illuminate\Image\ImageManager;
 
 /**
  * @method static \Illuminate\Contracts\Image\Driver driver(string|null $name = null)
- * @method static \Illuminate\Foundation\Image\Image fromBytes(string $contents)
- * @method static \Illuminate\Foundation\Image\Image fromPath(string $path)
- * @method static \Illuminate\Foundation\Image\Image fromUrl(string $url)
- * @method static \Illuminate\Foundation\Image\Image fromBase64(string $base64)
+ * @method static \Illuminate\Image\Image fromBytes(string $contents)
+ * @method static \Illuminate\Image\Image fromPath(string $path)
+ * @method static \Illuminate\Image\Image fromUrl(string $url)
+ * @method static \Illuminate\Image\Image fromBase64(string $base64)
  * @method static string getDefaultDriver()
- * @method static \Illuminate\Foundation\Image\ImageManager extend(string $driver, \Closure $callback)
+ * @method static \Illuminate\Image\ImageManager extend(string $driver, \Closure $callback)
  * @method static void pruneOrphaned(string|null $driver = null)
  *
  * @see ImageManager
